Se crearon los servicios de administrador y del cliente , con todos los metodos , queda faltando los servicios del admin de Teatro

+++ negocio/src/main/java/co/edu/uniquindio/unicine/dto/InformacionCompraDTO.java
@@ -15,6 +15,6 @@ public class InformacionCompraDTO {
     private Double precioTotal;
     private LocalDateTime fecha;
     private Funcion funcion;
-    private Integer precioEntradas;
-    private Integer precioConfiteria;
+    private Double precioEntradas;
+    private Double precioConfiteria;
 }
